pythonStuff/requirements.txt
# Core testing framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout>=2.2.0

# Browser automation
playwright==1.40.0

# Reporting
allure-pytest==2.13.2

# Environment and configuration
python-dotenv==1.0.0

# Utilities
Pillow==10.1.0

# Optional: API testing (if you need backend validation)
# requests>=2.31.0

# Optional: Data validation
# jsonschema>=4.21.1
